[Fix] #8 - 일정 인덱스 추가
